remove redundant project of MataCore
