+++ .gitignore
@@ -5,13 +5,9 @@
 /.vs
 /Mata/bin
 /Mata/obj
-/MataCore/bin
-/MataCore/obj
-/MataCore/project.lock.json
+/Mata/project.lock.json
 /Mata.Test/bin
 /Mata.Test/obj
-/MataCore.Test/bin
-/MataCore.Test/obj
-/MataCore.Test/project.lock.json
+/Mata.Test/project.lock.json
 /MataPackage/bin
 /MataPackage/obj
