--- Mata.Test/packages.config
@@ -1,12 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="NSubstitute" version="1.9.2.0" targetFramework="net461" />
-  <package id="StyleCop.Analyzers" version="1.0.0" targetFramework="net461" developmentDependency="true" />
-  <package id="xunit" version="2.1.0" targetFramework="net461" />
-  <package id="xunit.abstractions" version="2.0.0" targetFramework="net461" />
-  <package id="xunit.assert" version="2.1.0" targetFramework="net461" />
-  <package id="xunit.core" version="2.1.0" targetFramework="net461" />
-  <package id="xunit.extensibility.core" version="2.1.0" targetFramework="net461" />
-  <package id="xunit.extensibility.execution" version="2.1.0" targetFramework="net461" />
-  <package id="xunit.runner.visualstudio" version="2.1.0" targetFramework="net461" />
-</packages>
