+++ Mata.Test/project.json
@@ -4,7 +4,7 @@
   "dependencies": {
     "xunit": "2.2.0-beta2-build3300",
     "NSubstitute": "2.0.0-rc",
-    "MataCore": "1.0.0-*",
+    "Mata": "1.0.0-*",
     "dotnet-test-xunit": "2.2.0-preview2-build1029"
   },
   "frameworks": {
@@ -14,10 +14,9 @@
           "type": "platform",
           "version": "1.0.0"
         }
-      },
-      "imports": [
-        "netstandard1.6"
-      ]
-    }
+      }
+    },
+
+    "net461":  {} 
   }
 }
