+++ Mata.sln
@@ -3,8 +3,6 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
 VisualStudioVersion = 14.0.25420.1
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mata", "Mata\Mata.csproj", "{3BD785C9-A969-485D-A3C4-CC1C5BAD8FEC}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{64EC2F25-489D-4D6F-A071-2A2E41B2A165}"
 	ProjectSection(SolutionItems) = preProject
 		.gitignore = .gitignore
@@ -15,11 +13,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		VersionInfo.cs = VersionInfo.cs
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mata.Test", "Mata.Test\Mata.Test.csproj", "{BF7498D3-9481-4F69-8B1F-E1F2467E8244}"
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Mata", "Mata\Mata.xproj", "{8D93C2B8-25A7-4CC9-90F9-DEE1BBF81737}"
 EndProject
-Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "MataCore", "MataCore\MataCore.xproj", "{798A8B29-496C-41F9-A9FD-843FF9CB4F87}"
-EndProject
-Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "MataCore.Test", "MataCore.Test\MataCore.Test.xproj", "{99F7F90A-4428-4EED-80AE-A9D783B68D5A}"
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Mata.Test", "Mata.Test\Mata.Test.xproj", "{0684AB2E-5858-424B-8997-C787C3F23DFA}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -29,38 +25,22 @@ Global
 		Release-4.6.1|Any CPU = Release-4.6.1|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{3BD785C9-A969-485D-A3C4-CC1C5BAD8FEC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3BD785C9-A969-485D-A3C4-CC1C5BAD8FEC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3BD785C9-A969-485D-A3C4-CC1C5BAD8FEC}.Release|Any CPU.ActiveCfg = Release-4.6.1|Any CPU
-		{3BD785C9-A969-485D-A3C4-CC1C5BAD8FEC}.Release|Any CPU.Build.0 = Release-4.6.1|Any CPU
-		{3BD785C9-A969-485D-A3C4-CC1C5BAD8FEC}.Release-4.5.2|Any CPU.ActiveCfg = Release-4.5.2|Any CPU
-		{3BD785C9-A969-485D-A3C4-CC1C5BAD8FEC}.Release-4.5.2|Any CPU.Build.0 = Release-4.5.2|Any CPU
-		{3BD785C9-A969-485D-A3C4-CC1C5BAD8FEC}.Release-4.6.1|Any CPU.ActiveCfg = Release-4.6.1|Any CPU
-		{3BD785C9-A969-485D-A3C4-CC1C5BAD8FEC}.Release-4.6.1|Any CPU.Build.0 = Release-4.6.1|Any CPU
-		{BF7498D3-9481-4F69-8B1F-E1F2467E8244}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{BF7498D3-9481-4F69-8B1F-E1F2467E8244}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{BF7498D3-9481-4F69-8B1F-E1F2467E8244}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{BF7498D3-9481-4F69-8B1F-E1F2467E8244}.Release|Any CPU.Build.0 = Release|Any CPU
-		{BF7498D3-9481-4F69-8B1F-E1F2467E8244}.Release-4.5.2|Any CPU.ActiveCfg = Release|Any CPU
-		{BF7498D3-9481-4F69-8B1F-E1F2467E8244}.Release-4.5.2|Any CPU.Build.0 = Release|Any CPU
-		{BF7498D3-9481-4F69-8B1F-E1F2467E8244}.Release-4.6.1|Any CPU.ActiveCfg = Release|Any CPU
-		{BF7498D3-9481-4F69-8B1F-E1F2467E8244}.Release-4.6.1|Any CPU.Build.0 = Release|Any CPU
-		{798A8B29-496C-41F9-A9FD-843FF9CB4F87}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{798A8B29-496C-41F9-A9FD-843FF9CB4F87}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{798A8B29-496C-41F9-A9FD-843FF9CB4F87}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{798A8B29-496C-41F9-A9FD-843FF9CB4F87}.Release|Any CPU.Build.0 = Release|Any CPU
-		{798A8B29-496C-41F9-A9FD-843FF9CB4F87}.Release-4.5.2|Any CPU.ActiveCfg = Release|Any CPU
-		{798A8B29-496C-41F9-A9FD-843FF9CB4F87}.Release-4.5.2|Any CPU.Build.0 = Release|Any CPU
-		{798A8B29-496C-41F9-A9FD-843FF9CB4F87}.Release-4.6.1|Any CPU.ActiveCfg = Release|Any CPU
-		{798A8B29-496C-41F9-A9FD-843FF9CB4F87}.Release-4.6.1|Any CPU.Build.0 = Release|Any CPU
-		{99F7F90A-4428-4EED-80AE-A9D783B68D5A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{99F7F90A-4428-4EED-80AE-A9D783B68D5A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{99F7F90A-4428-4EED-80AE-A9D783B68D5A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{99F7F90A-4428-4EED-80AE-A9D783B68D5A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{99F7F90A-4428-4EED-80AE-A9D783B68D5A}.Release-4.5.2|Any CPU.ActiveCfg = Release|Any CPU
-		{99F7F90A-4428-4EED-80AE-A9D783B68D5A}.Release-4.5.2|Any CPU.Build.0 = Release|Any CPU
-		{99F7F90A-4428-4EED-80AE-A9D783B68D5A}.Release-4.6.1|Any CPU.ActiveCfg = Release|Any CPU
-		{99F7F90A-4428-4EED-80AE-A9D783B68D5A}.Release-4.6.1|Any CPU.Build.0 = Release|Any CPU
+		{8D93C2B8-25A7-4CC9-90F9-DEE1BBF81737}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8D93C2B8-25A7-4CC9-90F9-DEE1BBF81737}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8D93C2B8-25A7-4CC9-90F9-DEE1BBF81737}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8D93C2B8-25A7-4CC9-90F9-DEE1BBF81737}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8D93C2B8-25A7-4CC9-90F9-DEE1BBF81737}.Release-4.5.2|Any CPU.ActiveCfg = Release|Any CPU
+		{8D93C2B8-25A7-4CC9-90F9-DEE1BBF81737}.Release-4.5.2|Any CPU.Build.0 = Release|Any CPU
+		{8D93C2B8-25A7-4CC9-90F9-DEE1BBF81737}.Release-4.6.1|Any CPU.ActiveCfg = Release|Any CPU
+		{8D93C2B8-25A7-4CC9-90F9-DEE1BBF81737}.Release-4.6.1|Any CPU.Build.0 = Release|Any CPU
+		{0684AB2E-5858-424B-8997-C787C3F23DFA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0684AB2E-5858-424B-8997-C787C3F23DFA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0684AB2E-5858-424B-8997-C787C3F23DFA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0684AB2E-5858-424B-8997-C787C3F23DFA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0684AB2E-5858-424B-8997-C787C3F23DFA}.Release-4.5.2|Any CPU.ActiveCfg = Release|Any CPU
+		{0684AB2E-5858-424B-8997-C787C3F23DFA}.Release-4.5.2|Any CPU.Build.0 = Release|Any CPU
+		{0684AB2E-5858-424B-8997-C787C3F23DFA}.Release-4.6.1|Any CPU.ActiveCfg = Release|Any CPU
+		{0684AB2E-5858-424B-8997-C787C3F23DFA}.Release-4.6.1|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
