+++ Mata/Emit/MapEmitDebugInfo.cs
@@ -4,7 +4,9 @@
 
 namespace Mata.Emit
 {
+#if !NETSTANDARD1_6
     using System.Diagnostics.SymbolStore;
+#endif
     using System.IO;
     using System.Reflection;
     using System.Reflection.Emit;
@@ -22,8 +24,9 @@ namespace Mata.Emit
 
         private int line = 1;
 
+#if !NETSTANDARD1_6
         private ISymbolDocumentWriter symbolDocumentWriter;
-
+#endif
         public MapEmitDebugInfo(string typeNamespace, string typeName, string className)
         {
             this.typeNamespace = typeNamespace;
@@ -33,7 +36,9 @@ namespace Mata.Emit
 
         public void CreateDebugSymbols()
         {
+#if !NETSTANDARD1_6
             this.symbolDocumentWriter = MapEmitAssembly.CreateDocumentWriter(this.GetDebugSymbolName());
+#endif
             this.debugBuilder = new StringBuilder();
         }
 
@@ -143,7 +148,9 @@ namespace Mata.Emit
 
         public void SetLocalVariableName(LocalBuilder parametersLocal)
         {
+#if !NETSTANDARD1_6
             parametersLocal.SetLocalSymInfo("parameters");
+#endif
         }
 
         private static string GetDefaultValue(FieldMapDefinition field)
@@ -175,12 +182,14 @@ namespace Mata.Emit
 
         private void AddDebugSymbolLine(ILGenerator code, string lineText)
         {
+#if !NETSTANDARD1_6
             if (this.symbolDocumentWriter == null)
             {
                 return;
             }
 
             code.MarkSequencePoint(this.symbolDocumentWriter, this.line, 1, this.line, lineText.Length);
+#endif
             this.AddDebugLine(lineText);
         }
