--- Mata/Mata.nuspec
@@ -1,23 +1,0 @@
-<?xml version="1.0"?>
-<package xmlns="http://schemas.microsoft.com/packaging/2011/10/nuspec.xsd">
-  <metadata>
-    <id>Mata</id>
-    <version>0.9.0-alpha</version>
-    <title>Mata Data to POCO Mapping</title>
-    <authors>cleve.littlefield</authors>
-    <owners>cleve.littlefield</owners>
-    <licenseUrl>http://www.apache.org/licenses/LICENSE-2.0.html</licenseUrl>
-    <projectUrl>https://github.com/cleverguy25/Mata/wiki</projectUrl>
-    <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <description>Easily map data from DbDataReader to Poco objects or Poco to DbCommand parameters.  You can use DbDataReader/DbCommand that you already know, and still have an easy to maintain transformation to/from POCO objects.</description>
-    <summary>Mata Data to POCO Mapping</summary>
-    <copyright>Copyright © Palador Open Source</copyright>
-    <tags>Mata SqlDataReader POCO Micro ORM</tags>
-  </metadata>
-  <files>
-    <file src="bin\Release\net461\Mata.dll" target="lib\net461\Mata.dll" exclude="" />
-    <file src="bin\Release\net461\Mata.pdb" target="lib\net461\Mata.pdb" exclude="" />
-    <file src="bin\Release\net452\Mata.dll" target="lib\net452\Mata.dll" exclude="" />
-    <file src="bin\Release\net452\Mata.pdb" target="lib\net452\Mata.pdb" exclude="" />
-  </files>
-</package>
