+++ Mata/Properties/AssemblyInfo.cs
@@ -1,18 +1,14 @@
-﻿// <copyright file="AssemblyInfo.cs" company="Palador Open Source">
-// Copyright (c) Palador Open Source. All rights reserved. See License.txt in the project root for license information.
-// </copyright>
-
-using System.Reflection;
+﻿using System.Reflection;
+using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
 // General Information about an assembly is controlled through the following
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("Mata")]
-[assembly: AssemblyDescription("Mata Data to POCO Mapping")]
 [assembly: AssemblyConfiguration("")]
+[assembly: AssemblyCompany("")]
+[assembly: AssemblyProduct("Mata")]
 [assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
 
 // Setting ComVisible to false makes the types in this assembly not visible
 // to COM components.  If you need to access a type in this assembly from
@@ -20,4 +16,4 @@ using System.Runtime.InteropServices;
 [assembly: ComVisible(false)]
 
 // The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("3bd785c9-a969-485d-a3c4-cc1c5bad8fec")]
+[assembly: Guid("8d93c2b8-25a7-4cc9-90f9-dee1bbf81737")]
