+++ Mata/TypeExtensions.cs
@@ -5,6 +5,7 @@
 namespace Mata
 {
     using System;
+    using System.Reflection;
 
     public static class TypeExtensions
     {
@@ -20,8 +21,13 @@ namespace Mata
 
         public static bool IsNullableType(this Type propertyType)
         {
+#if NETSTANDARD1_6
+            return propertyType.GetTypeInfo().IsGenericType &&
+                   propertyType.GetGenericTypeDefinition() == typeof(Nullable<>);
+#else
             return propertyType.IsGenericType &&
                    propertyType.GetGenericTypeDefinition() == typeof(Nullable<>);
+#endif
         }
     }
 }
