--- Mata/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="NuGet.CommandLine" version="3.3.0" targetFramework="net461" />
-  <package id="StyleCop.Analyzers" version="1.0.0" targetFramework="net461" developmentDependency="true" />
-</packages>
