--- MataCore/project.json
@@ -1,21 +1,0 @@
-{
-  "version": "1.0.0-*",
-
-  "dependencies": {
-    "NETStandard.Library": "1.6.0",
-    "System.Data.Common": "4.1.0",
-    "System.Data.SqlClient": "4.1.0",
-    "System.Reflection.Emit": "4.0.1",
-    "System.Reflection.Emit.Lightweight": "4.0.1",
-    "System.Reflection.TypeExtensions": "4.1.0"
-  },
-
-  "frameworks": {
-    "netstandard1.6": {
-      "imports": [
-        "dnxcore50",
-        "net461"
-      ]
-    }
-  }
-}
